# portal/src/static/openapi.yml
openapi: 3.0.3
info:
  title: Healthcare Services API
  description: This is our Healthcare Services API that let you develop healthcare applications effortlessly.
  version: "1.0"
servers:
  - url: http://localhost:8002
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
security:
  - ApiKeyAuth: []
...
paths:
  /api/patients:
    get:
      tags:
        - "Patient Management"
      description: Get all Patients records
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                [{"id": "665186afb873a548422f9b79",
                  "name": "Patient 1",
                  "contact_no": "0123456789"},
                  {"id": "665186dbb873a548422f9b7b",
                  "name": "Patient 2",
                  "contact_no": "9876543210"}]
  
    post:
      tags:
        - "Patient Management"
      description: Add Patient Record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname:
                  type: string
                  example: <patient's firstname here>
                lastname:
                  type: string
                  example: <patient's lastname here>
                dob:
                  type: string
                  example: <dob here>
                gender:
                  type: string
                  enum:
                    - MALE
                    - FEMALE
                blood_group:
                  type: string
                  enum:
                    - A+
                    - A-
                    - B+
                    - B-
                    - O+
                    - O-
                    - AB+
                    - AB-
                contact_no:
                  type: string
                  example: <contact_no>
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Patient has been Added Successfully!
                  patient_name:
                    type: string
                    example: Patient 1
  /api/patients/{patient_id}:
    get:
      tags:
        - "Patient Management" 
      description: Get a Specific Patient record
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: 665186afb873a548422f9b79
      responses:
        200:
          description: Succcess
          content:
            application/json:
              example:
                {"_id": "665186afb873a548422f9b79",
                  "firstname": "patient",
                  "lastname": "1",
                  "age": "38",
                  "blood_group": "A+",
                  "contact_no": "0123456789",
                  "dob": "1985-08-14 00:00:00",
                  "gender": "MALE",
                  "uid": "6651862de66ba56c4dd11a9d"
                  }

    put:
      tags:
        - "Patient Management"
      description: Update a specific Patient record
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: 665186afb873a548422f9b79
      requestBody:
        required: True
        content:
          application/json:
            schema:
              type: object
              properties:
                field_you_want_to_update:
                  type: string
                  example: updated value
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Record updated Successfully!
                  patient_id:
                    type: string
                    example: 665186afb873a548422f9b79

    delete:
      tags:
        - "Patient Management"
      description: Delete a specific Patient record
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: 665186afb873a548422f9b79
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                message: Record Deleted Successfully!
...